enc dec with date
todo enc txt fiel

+++ main.py
@@ -1,13 +1,33 @@
 import sys
 import os
+import datetime
 from cryptography.fernet import Fernet
 
 
 def rmFile(path):
     try:
-        os.remove(path)
+        if os.path.exists(path):
+            os.remove(path)
+        else:
+            print("Can't delete none existing file")
     except Exception as e:
         print(e)
+
+
+def getCurrentTime():
+    try:
+        import time
+        import ntplib
+        import datetime
+
+        client = ntplib.NTPClient()
+        response = client.request('pool.ntp.org')
+        x = time.localtime(response.tx_time)
+        a = datetime.datetime(x.tm_year, x.tm_mon, x.tm_mday, x.tm_hour, x.tm_min, x.tm_sec)
+        print(a)
+        return a
+    except:
+        print('Could not sync with time server.')
 
 
 def addENC2File(path, add="ENC"):
@@ -29,10 +49,62 @@ def loadKey(key=b'VZPTDBB4K9dVyXUGdQBYC28fiXDbUjVaFEQUnGhOeY8='):
     return key, Fernet(key)
 
 
+def decUntil(path, key=""):
+    if not os.path.exists(path):
+        raise Exception("Time file does not exist")
+
+    file = open("encDate.txt", 'rt')
+    data = file.read()
+    file.close()
+
+    startPath = data.find(path)
+    nr = startPath + len(path) + 1  # + 1 for the comma
+    deadline = datetime.datetime.fromtimestamp(int(data[nr:nr+10]))
+    print("epoch : ", int(data[nr:nr+10]))
+    print("deadline", deadline)
+
+    current = getCurrentTime()
+    if current:
+        if current > deadline:
+            print("decrypting")
+            dec(path)
+            newData = data[:startPath - 1] + data[nr + 12:]
+            file = open("encDate.txt", 'wt')
+            file.write(newData)
+            file.close()
+            exit()
+        raise Exception("deadline not yet reached")
+    raise Exception("decyption failed")
+
+
+
+
+def encUntil(path, sec):
+    currentT = getCurrentTime()
+    if not currentT:
+        print(currentT)
+        raise Exception("server time none")
+    deadline = currentT + datetime.timedelta(seconds=10)
+
+    try:
+        if not os.path.exists(path):
+            file = open("encDate.txt", 'wt')
+        else:
+            file = open("encDate.txt", 'at')
+
+        enc(path)
+
+        file.write("\n")
+        file.write(path + "," + deadline.timestamp().__str__())
+        file.close()
+
+    except Exception as e:
+        print(e)
+
+
 def enc(path):
-    if os.path.exists(addENC2File(path)):
-        print("file already exist, delete before regenerating new one", addENC2File(path))
-        exit()
+    if not os.path.exists(path):
+        raise Exception("file does not exist")
 
     (key, fkey) = loadKey()
 
@@ -43,7 +115,9 @@ def enc(path):
 
         encry = fkey.encrypt(bytes)
 
-        newFile = open(addENC2File(path), "wb")
+        rmFile(path)
+
+        newFile = open(path, "wb")
         newFile.write(encry)
         newFile.close()
 
@@ -52,12 +126,11 @@ def enc(path):
         exit()
 
 
-def dec(path, key=b''):
+def dec(path, key=""):
     if not os.path.exists(path):
-        print("file does not exist")
-        exit()
+        raise Exception("file does not exist")
 
-    if key == b'' :
+    if not key:
         (key, fkey) = loadKey()
     else:
         (key, fkey) = loadKey(key)
@@ -69,13 +142,14 @@ def dec(path, key=b''):
 
         encry = fkey.decrypt(bytes)
 
-        newFile = open(renameFile(path, "name"), "wb")
+        rmFile(path)
+
+        newFile = open(path, "wb")
         newFile.write(encry)
         newFile.close()
 
     except Exception as e:
-        print(e)
-        exit()
+        raise Exception("decrypt failed",e)
 
 
 if __name__ == '__main__':
@@ -88,8 +162,19 @@ if __name__ == '__main__':
     elif arg1 == "enc":
         path = str(sys.argv[2])
         enc(path)
+    elif arg1 == "encUntil":
+        path = str(sys.argv[2])
+        AmountSec = int(sys.argv[3])
+        encUntil(path, AmountSec)
     elif arg1 == "dec":
         path = str(sys.argv[2])
         dec(path)
+    elif arg1 == "decUntil":
+        path = str(sys.argv[2])
+        decUntil(path)
     else:
-        print("wrong first arg")
+        print("usage: ")
+        print("enc C:\\Users\\ruben\\Desktop\\ImgCryp\\genericImag.png")
+        print("encUntil C:\\Users\\ruben\\Desktop\\ImgCryp\\genericImag.png 12")
+        print("dec c:\\Users\\ruben\\Desktop\\ImgCryp\\genericImag.png")
+        print("decUntil c:\\Users\\ruben\\Desktop\\ImgCryp\\genericImag.png")
